Fixing Travis errors related to homebrew The GCC builds for macOS both failed due to brew-related errors. The GCC 5 build failed with messages:   Error: Your Homebrew is outdated. Please run `brew update`.   Error: Kernel.exit This later caused the build to error because gcc-5 couldn't be found. Fixed by having brew update.
The GCC 7 build did update automatically because I was trying out brew as a command. But it failed with message:
  Error: No available formula with the name "gcc7"
Looks like it's instead called 'gcc@7' so that's an easy fix.

Also converted the GCC 7 matrix entry to the 'addons' form because even though it's more verbose, it's what Travis recommends on their website.